Notable resolutions on top of the route/resume architecture:
- adopted dev's flattened announce::Update { path, broadcast } (the
  Announced event enum this branch trimmed was removed entirely on dev)
- broadcast state moved to dev's alive + kio::Shared + Requests model;
  the route watch pairs a state poll with alive.poll_closed and the
  spliced queue polls the Shared directly
- track fetches use dev's coalescing FetchState, wrapped in this
  branch's Plain/Spliced dispatch
- route serve tasks run on the connection driver's TaskSet per dev's
  caller-driven sessions instead of free-floating spawns
- libmoq link flags come from dev's native-libs/*.txt single source
  (which already carries ScreenCaptureKit), superseding the ad-hoc list
- Subscription.stale renamed to latency_max throughout
- the route reannounce py test keeps its first hop stable across the
  update, matching the first-hop publisher-identity rule

+++ go/wrapper/moq/origin.go
@@ -91,7 +91,7 @@ func (r *BroadcastRequest) Accept(broadcast *BroadcastProducer) error {
 }
 
 // Abort fails the request with an application error code.
-func (r *BroadcastRequest) Abort(errorCode int32) error {
+func (r *BroadcastRequest) Abort(errorCode uint16) error {
 	return r.inner.Abort(errorCode)
 }
